requirements.txt
PyOpenGL
PyOpenGL-accelerate
numpy

